requirements.txt
Flask==3.1.1
python-dotenv==1.1.0

# RAG stack
langchain==0.3.26
langchain-community==0.3.26
langchain-text-splitters==0.3.11
langchain-pinecone==0.2.13
google-generativeai==0.8.6
sentence-transformers==4.1.0
pypdf==5.8.0

